reorganizing the code
1. conditionaal formating rule in Sample Map now has a try/catch
2. method openWikdShell to WSE lib, so it can be called from other scripts
3. WikdExtension class in its own file
